Extract formatting and advertisement handling logic outside of the GetSensorValuesUseCase file.
