+++ domain/src/main/java/com/aconno/acnsensa/domain/format/AdvertisementFormat.kt
@@ -1,0 +1,10 @@
+package com.aconno.acnsensa.domain.format
+
+/**
+ * @author aconno
+ */
+interface AdvertisementFormat {
+    fun getFormat(): Map<String, ByteFormat>
+    fun getRequiredFormat(): List<Byte>
+    fun getMaskBytePositions(): List<Int>
+}
