+++ domain/src/main/java/com/aconno/acnsensa/domain/format/ByteFormat.kt
@@ -1,0 +1,11 @@
+package com.aconno.acnsensa.domain.format
+
+/**
+ * @author aconno
+ */
+data class ByteFormat(
+    val startIndexInclusive: Int,
+    val endIndexExclusive: Int,
+    val isReversed: Boolean,
+    val targetType: String
+)
