+++ domain/src/main/java/com/aconno/acnsensa/domain/format/SupportedTypes.kt
@@ -1,0 +1,10 @@
+package com.aconno.acnsensa.domain.format
+
+/**
+ * @author aconno
+ */
+object SupportedTypes {
+    const val FLOAT = "Float"
+    const val SHORT = "Short"
+    const val BYTE = "Byte"
+}
